Move coin supply to fa processor
